tests/test_requirements.txt
django
djangorestframework==3.2.4
mysql-python
-e git+https://nl-bot:locum123@github.com/networklocum/cereal@master#egg=cereal